♻️ refactor : lobbySession 변수는 lobby.session.js 에서 가져올 수 있도록 코드 리팩토링
ref: #27
