+++ server/src/sessions/lobby.session.js
@@ -1,8 +1,12 @@
 import Lobby from '../classes/models/lobby.class.js';
-import { setLobbySession } from './session.js';
+import { lobbySession, setLobbySession } from './session.js';
 
 export const createLobbySession = () => {
   const session = new Lobby();
   setLobbySession(session);
   return session;
+};
+
+export const getLobbySession = () => {
+  return lobbySession;
 };
