+++ server/src/sessions/session.js
@@ -1,12 +1,8 @@
 export const userSessions = [];
 export const gameSessions = [];
 export const matchMakingSessions = [];
-let lobbySession;
+export let lobbySession;
 
 export const setLobbySession = (session) => {
   lobbySession = session;
-};
-
-export const getLobbySession = () => {
-  return lobbySession;
 };
